# android/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Phos API",
    "description": "AI-powered photo/video manager",
    "license": {
      "name": "MIT"
    },
    "version": "dev"
  },
  "paths": {
...
    "/api/people/{id}/browse": {
      "get": {
        "tags": [
          "people"
        ],
        "summary": "Get person browse graph",
        "description": "Returns a complete person browse graph with all shots and their file variants in a single response. Designed for offline-first browsing clients. The reserved id `unsorted` returns the shots that belong to nobody yet.",
        "operationId": "get_person_browse",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Person ID, or `unsorted` for shots with no person",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Person browse graph",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonBrowseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Person not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/settings/webdav": {
      "get": {
        "tags": [
          "Settings"
        ],
        "operationId": "get_webdav_settings",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebDavSettings"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "operationId": "set_webdav_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebDavCredentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Credentials saved"
          }
        }
      },
      "delete": {
        "tags": [
          "Settings"
        ],
        "operationId": "delete_webdav_settings",
        "responses": {
          "200": {
            "description": "WebDAV disabled"
          }
        }
      }
    },
...
    "/api/shots/{id}": {
      "get": {
        "tags": [
          "shots"
        ],
        "summary": "Get shot details",
        "description": "Retrieve full details for a single shot including all associated files, detected faces, and person assignments.",
        "operationId": "get_shot_detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Shot ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Shot detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShotDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Shot not found"
          }
        }
      },
      "put": {
        "tags": [
          "shots"
        ],
        "summary": "Update a shot",
        "description": "Update shot metadata such as the primary person assignment and review status.",
        "operationId": "update_shot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Shot ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShotPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Shot updated successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Shot not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "shots"
        ],
        "summary": "Delete a shot",
        "description": "Delete a shot and all its associated files from disk and database. This action is irreversible.",
        "operationId": "delete_shot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Shot ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Shot deleted successfully"
          },
          "404": {
            "description": "Shot not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlsoContainsPerson": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "BrowseFileDetail": {
        "type": "object",
        "required": [
          "id",
          "is_original",
          "thumbnail_url"
        ],
        "properties": {
          "file_size": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "is_original": {
            "type": "boolean"
          },
          "mime_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "thumbnail_url": {
            "type": "string"
          }
        }
      },
      "BrowseShotDetail": {
        "type": "object",
        "required": [
          "id",
          "files"
        ],
        "properties": {
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrowseFileDetail"
            }
          },
          "id": {
            "type": "string"
          },
          "review_status": {
            "type": [
              "string",
              "null"
            ]
          },
          "timestamp": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "FaceDetail": {
        "type": "object",
        "required": [
          "id",
          "file_id",
          "box_x1",
          "box_y1",
          "box_x2",
          "box_y2"
        ],
        "properties": {
          "box_x1": {
            "type": "number",
            "format": "float"
          },
          "box_x2": {
            "type": "number",
            "format": "float"
          },
          "box_y1": {
            "type": "number",
            "format": "float"
          },
          "box_y2": {
            "type": "number",
            "format": "float"
          },
          "file_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "person_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "person_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "FileDetail": {
        "type": "object",
        "required": [
          "id",
          "path",
          "is_original",
          "thumbnail_url"
        ],
        "properties": {
          "duration_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "file_size": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "is_original": {
            "type": "boolean"
          },
          "mime_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "path": {
            "type": "string"
          },
          "thumbnail_url": {
            "type": "string"
          },
          "width": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
      "PersonBrowseResponse": {
        "type": "object",
        "required": [
          "person",
          "shots"
        ],
        "properties": {
          "person": {
            "$ref": "#/components/schemas/PersonMeta"
          },
          "shots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrowseShotDetail"
            }
          }
        }
      },
...
      "PersonMeta": {
        "type": "object",
        "description": "Browse a person: returns person metadata + all shots + all files per shot.\nSolves the N+1 problem for the mobile client.",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ShotDetailResponse": {
        "type": "object",
        "required": [
          "id",
          "files",
          "faces",
          "also_contains"
        ],
        "properties": {
          "also_contains": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlsoContainsPerson"
            }
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "faces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FaceDetail"
            }
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileDetail"
            }
          },
          "folder_number": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "next_shot_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "prev_shot_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "primary_person_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "primary_person_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "review_status": {
            "type": [
              "string",
              "null"
            ]
          },
          "timestamp": {
            "type": [
              "string",
              "null"
            ]
          },
          "width": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
      "UpdateShotPayload": {
        "type": "object",
        "description": "PUT /api/shots/:id - update primary_person_id and/or review_status.\nWhen primary_person changes, assign new folder_number (MAX+1 for that person).",
        "properties": {
          "primary_person_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "review_status": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "WebDavCredentials": {
        "type": "object",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
      "WebDavSettings": {
        "type": "object",
        "required": [
          "enabled",
          "username"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "username": {
            "type": "string",
            "description": "The WebDAV username (OIDC sub in multi-user mode, \"phos\" in single-user)."
          }
        }
      }
    },
    "securitySchemes": {
      "session_cookie": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Session JWT set as HttpOnly cookie (phos_session) after OIDC login"
      }
    }
  },
  "security": [
    {
      "session_cookie": []
    }
  ]
}